backend/requirements/prod.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.2
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
factory_boy==3.3.3
Faker==40.1.2
flake8==7.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mccabe==0.7.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
